Re-organizing, Separating Jack
I changed and moves a lot in order to separate jack to his own file.
Some of the things I did to this include:

- I made the camera global, so jack.cpp can alter it. This will be
changed later when I implement a separate camera.cpp to control the
camera.
- I made a couple functions global and defined them in a new file
"FunctionLib.h" and I initialized them in main. This we may want to
keep.
- Obviiously moved all of jacks functions to Jack.cpp
- Added functions to jack.cpp and jack.h in order to control jacks
movement.

Lost some functionality for the moment, not the following:
- The S key to look down no longer works. Also should be fixed with the
addition of camera.cpp
- The controls atm do not account for multiple jump prevention and mess
up a little bit on multiple key pressed. I am working on the fixed, but
it is all involved with the fix for the animation code.
- As mentioned, the animation still has the same errors as before.
- Also does not include the changes Steve made to include a cursor and
etc.

What needs to be done is to fix the above and somehow implement changes
made to the master code.