Fixed Animation Handler
The animations play as they should. Only issue is a jitter when the
player hits the ground.

Possible solution may ne to investigate the move function and what
happens when colliding. Now it seems to just move the player back the
last increment of velocity, when it should move the player Just on top
of the platform.

Maybe grab the y value of the platform and set the player relative to
that.