Separating platforms, re-thinking collisions
Platforms now have separate files and class.
-Platforms are not visible yet
-All other functions are stable except for s key function
-Animation glitch showed when making these changes, it was not holding
the jumping and falling frames. Fixed by changing the frame number from
4 to 3 for the conditional statement.

Still to do:
-Fix the s key function while re-designing how the camera works
-Fix controls, they mess up with multiple button presses
-Change how platforms are made, made a collision list, decide how that
will work so we can pass the y position of the platform to the player
when colliding so the player can just be shifted on top of the platform.