Update with screen changes
No functionality change, but screen now works as an member of the Window
class. I felt it necessary to upload this proof of concept. Still need
to get background to function the same way so platforms can be apply to
it.

The way to access the screen is shown by The use of it in the
Jack.Show() function, which now takes an object of type Window by
reference.