Platforms now show, found new errors.
Changes Since Last Commit
-Platforms now show and are a part of the window class.
-Now background and screen can be accessed via functions.
-Load files works differently, and needs to be updated for jack.
Essentially now load_files gets passed the window, then it will run the
load_files function for every other object class we are going to use.
-Other changes may have been made to achieve functionality

What needs to be addressed still:
-Clean_up function should be updated after all of these changes so it
will eventually function much like load_files does.
- For some reason, the background only showed if the surface was
declared in Global.h This is very inconvenient when it comes to
expanding the code. We need it so background can be be completely
separated to the window class files. Then do the same for jack and the
platforms.
- Still not yet fixed the camera shift functionality, it seems like a
lower priority at the moment.

Possible New Errors:
-It appears we may need to clear surfaces after we apply them for each
frame. I noticed when I couldn't get the background to show, every frame
for Jack was staying on the screen, creating a trailing effect.
- This error can be seen by commenting out the line in main for showing
the background. It seems to me that we need to another clean up function
or something so we clear the last frame that was on the screen in the
last frame before we put a new one on.
- The only reason why we did not notice before is because we were
showing the background again before showing jack again for each frame,
so the background covered the old jack.

On the bright side this may have revealed how to achieve a cool trailing
effect.