Started the re-build
I have a stable re-build of the collision system. It is a simple proof
of concept with the basics set up to implement polymorphism with the
Platform class. Next will Implement triangle ramps to see if it
functions.

There should not be any improvement yet as to the functionality of
colliding, but this will be fixed soon.

Still cannot seem to fix my gitignore file.