Identified an error. Making a note of it.
-I reverted to the simple method of collision checking, where no matter
how you collide with a platform, you go on top of it. This was in order
to identify errors and how best to approach the intended changes.

-There is an error in the way collisions are set up. Because the probe
for the onGround flag is smaller than the width of the player, when
going off edges, there is a zone in which falling actions take place,
but so does the collision actions, so the result is a falling animation
and Y acceleration, but the player is still vertically set to be on top
of the platform.

-possible solutions include changing the size of the probe hit box for
the onGround flag, and changing the conditions on which the collision
actions are set.