Upload before a change of plans
- mostly fixed position issue for side collisions.
- noticed more errors, so re-thinking how collisions work. uploading
before i start changing things.
Plans include:
- making it so there is only a response call to the jack object when
there is a collision of the main bounding box.
- Make the system only check for side and ground collisions if the flag
is set. Possibly have 2 flags, one in the platform, and one in the
player. Either that or have the player with a pointer to the platform as
I did before.
Make the sensitivity for side collisions dependant on positions instead
of a collection of collide checks.
-perhaps have the checks to unset the flags more literal, like side
collisions are now.